Custom Path Editor
Created a custom path editor to define the birds path.
Gives control to designers to choose which waypoints and the flight times between thye waypoints.
Can be set to a loop or a curve with an end.
TO DO: Create function to say when a waypoint has been reached so that the pathFollower can stop and restart at the waypoints.